Added Equipment and Apparel powers as well as removed extra files.
Equipment and Apparel can now have powers on them just like Pawns. I
created GenericCompAbilityUser to place these in on the Pawn. If
the AbilityUserClass tag is specified in the XML, it can be specific to
a specific CompAbilityUser (i.e. only Wizards can use this power).

This is a crappy commit with debugging spew so I will remove that in a
follow-up commit. I just want to get this code in.

I also removed extra bin/obj directories and dlls, since JecTools now
depends on the HugsLib Mod to be installed. This will be a problem
during development for you Jecrell. I use a directory with all mods in
it which has a GlobalAssemblies folder where I put 0Harmony.dll,
   Assembly-CSharp.dll, HugsLib.dll and UnityEngine.dll. This way I
   don’t have to track versions of all dlls and have incompatibilities,
   especially when they install.