Upgrade to latest RW nuget, remove obsolete Manifest.xml fields
The removed Manifest.xml fields are loadAfter/loadBefore, which are obsolete in RW 1.1 and causes lots of warnings to be logged.
Although they are still relevant for RW 1.0, it's currently not possible to multi-version About files like Manifest.xml,
and it's not worth supporting the few remaining RW 1.0 users that are still using Mod Manager to build new mod lists.