More CompOversizedWeapon optimizations, including better thing-on-ground detection
The previous thing-on-ground detection method of compEquippable.verbTracker.PrimaryVerb.CasterPawn turned out
to be unreliable and it was only calculated every 60 ticks for performance reasons.
The new method takes advantage of the thing's ParentHolder being Map if it's on the ground. Not only is this
a fast check (always calculated on demand), it doesn't require CompEquippable at all.
Also removed the FirstAttack property that never worked. This change breaks source compatibility, but AFAIK no mod
out in the wild programmatically uses GetEquippable/IsEquipped/FirstAttack.

Harmony patch code has also been both further optimized and structurally improved a bit.
Added a note that the DefaultGraphics drawSize workaround, one of the original raison d'etre for CompOversizedWeapon,
isn't actually needed in RW 1.2.2723+, but the workaround is kept for supporting older RW versions.